Update on Entities: id fields changed to Long.

+++ src/main/java/org/hubson404/miniSocialNetwork/model/ForwardBadge.java
@@ -14,7 +14,7 @@ public class ForwardBadge {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    int forwardId;
+    Long forwardId;
 
     @CreationTimestamp
     LocalDateTime createDate;
